add dnscrypt-wrapper.{c,h} and move wrapper implementation specific code there.
dnscrypt.{c,h} and cert.{c,h} are now the only files dealing with the
implementation of the protocol.
