+++ cert.h
@@ -2,6 +2,7 @@
 #define CERT_H
 
 #include <sodium.h>
+#include <time.h>
 #define CERT_MAGIC_CERT "DNSC"
 #define CERT_MAJOR_VERSION 1
 #define CERT_MINOR_VERSION 0
