+++ edns.c
@@ -1,4 +1,4 @@
-#include "dnscrypt.h"
+#include "dnscrypt-wrapper.h"
 
 static int
 _skip_name(const uint8_t *const dns_packet, const size_t dns_packet_len,
