+++ main.c
@@ -1,4 +1,4 @@
-#include "dnscrypt.h"
+#include "dnscrypt-wrapper.h"
 #include "argparse/argparse.h"
 #include "version.h"
 #include "pidfile.h"
