+++ pidfile.h
@@ -1,7 +1,7 @@
 #ifndef PIDFILE_H
 #define PIDFILE_H
 
-#include "dnscrypt.h"
+#include "dnscrypt-wrapper.h"
 
 int pidfile_create(const char *const pidfile);
