+++ tcp_request.h
@@ -1,7 +1,7 @@
 #ifndef TCP_REQUEST_H
 #define TCP_REQUEST_H
 
-#include "dnscrypt.h"
+#include "dnscrypt-wrapper.h"
 
 #define DNS_MAX_PACKET_SIZE_TCP (65535U + 2U)
